+++ experimental/packages/opentelemetry-sdk-node/src/start.ts
@@ -12,6 +12,7 @@ import {
   diag,
   DiagConsoleLogger,
   metrics,
+  trace,
   propagation,
 } from '@opentelemetry/api';
 import {
@@ -22,6 +23,8 @@ import {
   getPropagatorFromConfiguration,
   getResourceDetectorsFromConfiguration,
   getResourceFromConfiguration,
+  getSpanLimitsFromConfiguration,
+  getSpanProcessorsFromConfiguration,
 } from './utils';
 import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import type { SDKComponents, SDKOptions } from './types';
@@ -40,6 +43,7 @@ import {
 } from '@opentelemetry/resources';
 import { AsyncLocalStorageContextManager } from '@opentelemetry/context-async-hooks';
 import { ATTR_SERVICE_INSTANCE_ID } from './semconv';
+import { BasicTracerProvider } from '@opentelemetry/sdk-trace-base';
 
 /**
  * @experimental Function to start the OpenTelemetry Node SDK
@@ -71,6 +75,9 @@ export function startNodeSDK(sdkOptions: SDKOptions): {
   if (components.meterProvider) {
     metrics.setGlobalMeterProvider(components.meterProvider);
   }
+  if (components.tracerProvider) {
+    trace.setGlobalTracerProvider(components.tracerProvider);
+  }
   if (components.propagator) {
     propagation.setGlobalPropagator(components.propagator);
   }
@@ -82,6 +89,9 @@ export function startNodeSDK(sdkOptions: SDKOptions): {
     }
     if (components.meterProvider) {
       promises.push(components.meterProvider.shutdown());
+    }
+    if (components.tracerProvider) {
+      promises.push(components.tracerProvider.shutdown());
     }
     await Promise.all(promises);
   };
@@ -132,6 +142,25 @@ function create(
       views: meterViews ?? [],
     });
     components.meterProvider = meterProvider;
+  }
+
+  const spanProcessors = getSpanProcessorsFromConfiguration(config);
+  if (spanProcessors) {
+    const spanLimits = getSpanLimitsFromConfiguration(config);
+    // TODO (6506): support sampler configuration from config
+    const tracerProvider = new BasicTracerProvider({
+      resource,
+      spanProcessors,
+      spanLimits,
+      generalLimits: {
+        attributeValueLengthLimit:
+          config.attribute_limits?.attribute_value_length_limit,
+        attributeCountLimit: config.attribute_limits?.attribute_count_limit,
+      },
+      // TODO (6616): support idGenerator configuration from config
+      // TODO (6624): support for `meterProvider: components.meterProvider`
+    });
+    components.tracerProvider = tracerProvider;
   }
 
   return components;